+++ ui/src/app.tsx
@@ -52,7 +52,6 @@ import EditProfile from '@/profiles/EditProfile/EditProfile';
 import HeapDetail from '@/heap/HeapDetail';
 import groupsFavicon from '@/assets/groups.svg';
 import talkFavicon from '@/assets/talk.svg';
-import { usePendingGangsWithoutClaim } from '@/state/groups/groups';
 import GroupInvitesPrivacy from './groups/GroupAdmin/GroupInvitesPrivacy';
 import Notifications, { MainWrapper } from './notifications/Notifications';
 import ChatChannel from './chat/ChatChannel';
@@ -91,8 +90,8 @@ import SettingsDialog from './components/SettingsDialog';
 import { captureAnalyticsEvent } from './logic/analytics';
 import GroupChannel from './groups/GroupChannel';
 import PrivacyNotice from './groups/PrivacyNotice';
-import useAutoJoinLureInvites from './groups/autoJoinLureInvites';
 import ActivityModal, { ActivityChecker } from './components/ActivityModal';
+import LureAutojoiner from './groups/LureAutojoiner';
 
 const Grid = React.lazy(() => import('./components/Grid/grid'));
 const TileInfo = React.lazy(() => import('./components/Grid/tileinfo'));
@@ -280,14 +279,6 @@ function HomeRoute({ isMobile = true }: { isMobile: boolean }) {
   const navigate = useNavigate();
   const groups = queryClient.getQueryCache().find(['groups'])?.state.data;
   const isInGroups = groups !== undefined ? !_.isEmpty(groups) : true;
-  const pendingGangsWithoutClaim = usePendingGangsWithoutClaim();
-  const autojoin = useAutoJoinLureInvites();
-
-  useEffect(() => {
-    if (Object.keys(pendingGangsWithoutClaim).length) {
-      autojoin(pendingGangsWithoutClaim);
-    }
-  }, [pendingGangsWithoutClaim, autojoin]);
 
   useEffect(() => {
     if (!isInGroups && redirectToFind) {
@@ -737,6 +728,7 @@ function RoutedApp() {
           <Scheduler />
           {import.meta.env.DEV && <Eyrie />}
         </TooltipProvider>
+        <LureAutojoiner />
         <ReactQueryDevtools initialIsOpen={false} />
       </Router>
     </ErrorBoundary>
